requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
deprecation==2.1.0
exceptiongroup==1.2.2
extra-streamlit-components==0.1.71
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gotrue==2.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.13.3
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
postgrest==0.18.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
realtime==2.0.6
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
storage3==0.9.0
streamlit==1.40.0
streamlit-cookies-controller==0.0.4
StrEnum==0.4.15
supabase==2.10.0
supafunc==0.7.0
sympy==1.13.1
tenacity==9.0.0
toml==0.10.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.32
ultralytics-thop==2.0.11
urllib3==2.2.3
watchdog==5.0.3
websockets==13.1
yarl==1.17.1
